CLI: Move subscribe steps up the directory hierarchy
