+++ packages/frontity/src/scripts.ts
@@ -1,3 +1,3 @@
 export { default as create } from "./commands/create";
-export { default as subscribe } from "./steps/subscribe";
+export { default as subscribe } from "./steps";
 export * from "@frontity/core";
